AI/requirements.txt
flask==2.3.3
pandas==2.0.3
langchain-community==0.2.10
langchain-core==0.2.23
chromadb==0.4.15
sentence-transformers==2.2.2
openpyxl==3.1.2
werkzeug==2.3.7
flask-cors==4.0.0
python-dotenv==1.0.1 